+++ app/services/bot_context/commands/book.rb
@@ -8,6 +8,7 @@ module BotContext
         remove_book_command: 'commands.homebrew_context.remove_book',
         add_book_races_command: 'commands.homebrew_context.daggerheart.add_book_races',
         add_book_communities_command: 'commands.homebrew_context.daggerheart.add_book_communities',
+        add_book_transformations_command: 'commands.homebrew_context.daggerheart.add_book_transformations',
         add_book_subclasses_command: 'commands.homebrew_context.daggerheart.add_book_subclasses',
         add_book_items_command: 'commands.homebrew_context.daggerheart.add_book_items',
         copy_book_command: 'commands.homebrew_context.copy_book'
@@ -25,6 +26,7 @@ module BotContext
         when 'set' then set_book(*arguments, data, source)
         when 'addRace' then add_race(*arguments, data, source)
         when 'addCommunity' then add_community(*arguments, data, source)
+        when 'addTransformation' then add_transformation(*arguments, data, source)
         when 'addSubclass' then add_subclass(*arguments, data, source)
         when 'addItem' then add_item(*arguments, data, source)
         when 'export' then export_book(data, source)
@@ -100,6 +102,16 @@ module BotContext
         add_book_communities_command.call(user: data[:user], book: ::Homebrew::Book.find(object.info['id']), names: arguments)
         {
           type: 'add_community',
+          result: :ok,
+          errors: nil
+        }
+      end
+
+      def add_transformation(*arguments, data, source)
+        object = ActiveBotObject.find_by!(user: data[:user], source: source, object: 'book')
+        add_book_transformations_command.call(user: data[:user], book: ::Homebrew::Book.find(object.info['id']), names: arguments)
+        {
+          type: 'add_transformation',
           result: :ok,
           errors: nil
         }