+++ db/migrate/20251003190431_create_homebrew_transformations.rb
@@ -1,0 +1,12 @@
+class CreateHomebrewTransformations < ActiveRecord::Migration[8.0]
+  disable_ddl_transaction!
+
+  def change
+    create_table :daggerheart_homebrew_transformations, id: :uuid do |t|
+      t.uuid :user_id, null: false
+      t.string :name, null: false
+      t.timestamps
+    end
+    add_index :daggerheart_homebrew_transformations, :user_id, algorithm: :concurrently
+  end
+end