+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - shared homebrew books
+- transformations for Daggerheart characters
 
 ### Modified
 - setting damage and trait for Daggerjeart homebrew weapons